+++ src/clj_kondo/impl/config.clj
@@ -37,6 +37,7 @@
               :duplicate-require {:level :warning}
               :duplicate-field {:level :error}
               :duplicate-key-in-assoc {:level :warning}
+              :duplicate-key-in-dissoc {:level :warning}
               :missing-map-value {:level :error}
               :redefined-var {:level :warning}
               :var-same-name-except-case {:level :warning}